refactor(utils): split DataUtils tests for GH727
